ID will not be persistent after loading from file.
This will also make the gaps remove.

Reason:
If a person was deleted the id becomes vacant, but during loading the counter was increased with every person added. This makes the counter lag behind the last id, which in turn causes the id of new person added same as the last one.

Another solution:
Counter can be increased to the highest id received.